Monitor for fatal damage and reset the flag when it occurs.
The problem with monitoring for the player's death for resetting the flag appears to be that the player's inventory is wiped before the event gets to our script which means they never have the flag, even if they did when they died. Rather than monitor for the death of a player we instead monitor for damage taken and conclude ourselves if the damage they took is fatal.

Damage events are triggered before death events so the player's inventory should still be intact, allowing us to check if they have the flag just before they die and reset the flag. I also remove the flag from the player just in-case something goes wrong and the player does not die, as that would cause there to be two active flags for a team which would be bad.